- Creates new AssetRenderer component to avoid reloading asset repeatedly on page
